bet-check/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.5
python-dotenv==1.0.0
supabase==2.3.4
requests==2.31.0
httpx==0.25.0
numpy>=1.26.0
